\name{morph-package}

\alias{morph-package}

\alias{morph}

\docType{package}

\title{
\packageTitle{morph}
}

\description{
\packageDescription{morph}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{morph}
\packageIndices{morph}
This package handles the segmentation of binary voxel data, provided as a numeric 3D array, into its constituent morphological classes. Each voxel is assigned to one of nine mutually exclusive morpholocal classes: outside, mass, skin, crumb, circuit, antenna, bond, void-volume, or void.
}

\author{
\packageAuthor{morph}
Maintainer: \packageMaintainer{morph}
}

\references{
Remmel, T.K. (Accepted December 2021). Extending morphological pattern segmentation to 3D voxels. Landscape Ecology.
}

\keyword{ package }
\keyword{ manip }
\keyword{ iplot }

\seealso{
See Also \code{\link[reshape2]{melt}}
}

\examples{
# Need to provide some simple examples here.
}
